requirements.txt
crawl4ai>=0.8.0
playwright

sentence-transformers
numpy

psycopg2-binary

fastapi
uvicorn[standard]

python-dotenv